requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
coinbase==2.1.0
colorama==0.4.4
cycler==0.10.0
discord==1.7.3
discord.py==1.7.3
future==0.18.2
gTTS==2.2.3
HandTrackingModule==0.1
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.2
mediapipe==0.8.6
MouseInfo==0.1.3
mpyg321==1.3.0
multidict==5.1.0
numpy==1.21.0
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
pandas==1.3.1
pefile==2021.5.24
Pillow==8.3.1
protobuf==3.17.3
PyAutoGUI==0.9.53
pycryptodome==3.10.1
PyGetWindow==0.0.9
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-dateutil==2.8.2
python-dotenv==0.19.0
PyTweening==1.0.3
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.26.0
six==1.16.0
SpeechRecognition==3.8.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3